+++ snoozy-core/src/main/java/fi/jubic/snoozy/auth/TokenParser.java
@@ -4,5 +4,9 @@ import javax.servlet.http.HttpServletRequest;
 import java.util.Optional;
 
 public interface TokenParser {
+    /**
+     * Extract authentication token from a {@link HttpServletRequest}.
+     * @return An {@link Optional} of the token, or empty for HTTP 401.
+     */
     Optional<String> parse(HttpServletRequest servletRequest);
 }